# Silence ZDNS status updates with `make benchmark`